2025-03-11 14:39 by shun - 미지급금 처리 완료

+++ src/main/java/com/beauty1nside/accnut/dto/DealBookDTO.java
@@ -11,7 +11,7 @@ public class DealBookDTO {
 	String dealingsAccountBookCode;
 	String section;
 	String typesOfTransaction;
-	int amount;
+	Long amount;
 	String vatAlternative;
 	String dealingsContents;
 //	@DateTimeFormat(pattern = "yyyy-MM-dd")
